requirements.txt
pyyaml
pandas
numpy
anthropic
openai
httpx
pydantic
pydantic[email]
fastapi
uvicorn
edgartools
beautifulsoup4
lxml
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
